+++ test/vanilla/low-level/Expected/AcceptanceTests/MediaTypesLowLevel/mediatypeslowlevel/rest/__init__.py
@@ -11,6 +11,7 @@ from ._request_builders import build_analyze_body_no_accept_header_request
 from ._request_builders import build_content_type_with_encoding_request
 from ._request_builders import build_binary_body_with_two_content_types_request
 from ._request_builders import build_binary_body_with_three_content_types_request
+from ._request_builders import build_body_three_types_request
 from ._request_builders import build_put_text_and_json_body_request
 
 __all__ = [
@@ -19,5 +20,6 @@ __all__ = [
     "build_content_type_with_encoding_request",
     "build_binary_body_with_two_content_types_request",
     "build_binary_body_with_three_content_types_request",
+    "build_body_three_types_request",
     "build_put_text_and_json_body_request",
 ]